adjust dtype for training outputs

+++ sklearnBPMF/core/wrapper.py
@@ -163,12 +163,12 @@ class Wrapper(BaseEstimator):
 
         # Get test predictions
         predictions = self.trainSession.getTestPredictions()
-        predAvg = [p.pred_avg for p in predictions]
+        predAvg = [np.float32(p.pred_avg) for p in predictions]
         predCoord = [p.coords for p in predictions]
 
         if return_std:
             predVar = [np.var(p.pred_all) for p in predictions]
-            predStd = [np.sqrt(p) for p in predVar]
+            predStd = [np.float32(np.sqrt(p)) for p in predVar]
             return predAvg, predStd, predCoord
         else:
             return predAvg, predCoord
@@ -180,12 +180,12 @@ class Wrapper(BaseEstimator):
         predictor = self.trainSession.makePredictSession()
         p_all = predictor.predict_all()
         p_all_avg = sum(p_all) / len(p_all)
-        predAvg = list(p_all_avg[self.train_coords[0],self.train_coords[1]])
+        predAvg = list(p_all_avg[self.train_coords[0],self.train_coords[1]].astype(np.float32))
         predCoord = list(zip(*self.train_coords))
 
         if return_std:
             predVar = np.var([p[self.train_coords[0], self.train_coords[1]] for p in p_all],axis=0)
-            predStd = np.square(predVar).tolist()
+            predStd = np.square(predVar).astype(np.float32).tolist()
             return predAvg, predStd, predCoord
         else:
             return predAvg, predCoord
